requirements.txt
insightface>=0.7.3
onnxruntime>=1.17.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0

